+++ CHANGELOG.md
@@ -4,7 +4,10 @@ Change Log
 ## 0.8.0 (TBD)
 
 ### Added
-- ViewModel extension enabling to launch Coroutine on viewModelScope without writing the viewModelScope part.
+- `BindingRecyclerViewAdapter` to be used instead of classic adapters.
+- `bindRecyclerView` databinding function that binds `BindingRecyclerViewAdapter` into RecyclerView.
+- `bindViewPager2` databinding function that binds `BindingRecyclerViewAdapter` into ViewPager2.
+- ViewModel extension enabling to launch Coroutine on `viewModelScope` without writing the viewModelScope part.
 
 ### Changed
 - Gradle updated to [v6.4.1](https://docs.gradle.org/6.4.1/release-notes.html)